refactor: change file layout and update dependencies
- Move generate, merge, logging, and validate packages to internal/
- Update go.mod dependencies
