+++ sbomtool/go.work
@@ -1,4 +1,4 @@
-go 1.24.2
+go 1.24.3
 
 use .
