+++ sbomtool/internal/commands/generate/generate.go
