+++ sbomtool/internal/commands/generate/generate_test.go
