+++ sbomtool/internal/logging/logging_test.go
