+++ sbomtool/internal/validate/validate.go
