+++ sbomtool/internal/validate/validate_test.go
